Refactor wallet UI and update documentation
  - Fix invisible button visibility issues (New Wallet, Import Wallet, No Thanks
  buttons)
  - Improve button layout: Move Send button to right, Receive to left
  - Fix backup secret phrase functionality with proper error handling
  - Implement automatic server synchronization with periodic updates
  - Update README.md with comprehensive build instructions and tool versions
  - Add network error handling and retry mechanisms
  - Clean up deprecated code and improve overall stability

+++ lib/util/numberutil.dart
@@ -1,4 +1,4 @@
-// @dart=2.9
+
 
 // Dart imports:
 import 'dart:math';
@@ -27,7 +27,7 @@ class NumberUtil {
   ///
   static double truncateDecimal(Decimal input,
       {int digits = maxDecimalDigits}) {
-    return (input * Decimal.fromInt(pow(10, digits))).truncateToDouble() /
+    return (input * Decimal.fromInt(pow(10, digits).toInt())).truncateToDouble() /
         pow(10, digits);
   }
